+++ Ambassador/src/utopia/ambassador/rest/resource/task/TaskAccessTestNode.scala
@@ -13,6 +13,7 @@ import utopia.exodus.rest.util.AuthorizedContext
 import utopia.flow.datastructure.immutable.Model
 import utopia.flow.generic.ValueConversions._
 import utopia.flow.util.CollectionExtensions._
+import utopia.metropolis.model.cached.LanguageIds
 import utopia.metropolis.model.enumeration.ModelStyle.{Full, Simple}
 import utopia.nexus.http.Path
 import utopia.nexus.rest.LeafResource
@@ -57,8 +58,8 @@ case class TaskAccessTestNode(taskId: Int) extends LeafResource[AuthorizedContex
 					// Reads the descriptions of the missing scopes
 					val missingScopes = remainingRequiredScopes ++ remainingAlternativeGroups.values.flatten
 					val missingScopeIds = missingScopes.map { _.id }.toSet
-					val descriptions = DbScopeDescriptions(missingScopeIds)
-						.inLanguages(context.languageIdListFor(session.userId))
+					implicit val languageIds: LanguageIds = context.languageIdListFor(session.userId)
+					val descriptions = DbScopeDescriptions(missingScopeIds).inPreferredLanguages
 					// Reads service names
 					val services = DbAuthServices(missingScopes.map { _.serviceId }.toSet).pull