+++ Exodus/src/utopia/exodus/rest/resource/description/TasksNode.scala
@@ -5,6 +5,7 @@ import utopia.citadel.database.access.id.many.DbTaskIds
 import utopia.citadel.database.access.many.description.{DbDescriptionRoles, DbTaskDescriptions}
 import utopia.exodus.rest.util.AuthorizedContext
 import utopia.flow.generic.ValueConversions._
+import utopia.metropolis.model.cached.LanguageIds
 import utopia.metropolis.model.combined.organization.DescribedTask
 import utopia.metropolis.model.enumeration.ModelStyle.{Full, Simple}
 import utopia.nexus.http.Path
@@ -36,9 +37,9 @@ object TasksNode extends Resource[AuthorizedContext]
 	{
 		context.sessionKeyAuthorized { (session, connection) =>
 			implicit val c: Connection = connection
-			val languageIds = context.languageIdListFor(session.userId)
+			implicit val languageIds: LanguageIds = context.languageIdListFor(session.userId)
 			// Reads task descriptions
-			val descriptions = DbTaskDescriptions.inLanguages(languageIds)
+			val descriptions = DbTaskDescriptions.inPreferredLanguages
 			// Combines the descriptions with the tasks and returns them
 			val describedTasks = DbTaskIds.all.map { taskId => DescribedTask(taskId,
 				descriptions.getOrElse(taskId, Set()).toSet) }